se adiciona imagenes para background y sprites. Se modifica escenario y se modifica personaje principal y resorte

+++ objetodinamico.h
@@ -29,9 +29,9 @@ public:
     void MoveLeft();
     void MoveUp();
     void MoveDown();
-    double getPosx() ;
+    float getPosx() ;
     void setPosx(float newPosx);
-    double getPosy() ;
+    float getPosy() ;
     void setPosy(float newPosy);
 
     //Metodos abstractos
